yatm_security: Reworking programmer table prog_data
Originally it was a text box where players would type in a random string.

Part of the refactor for YATM is to move away from textboxes as much as possible for certain configurations.

Instead a 8x8 bitmap is used to declare the prog data (secret), this allows a 64 bit value via 8 bytes, each row is a byte.